Save local changes

+++ notes/storage.py
@@ -1,20 +1,22 @@
-from fastapi import Depends, HTTPException, Query
 from typing import Annotated
 from sqlalchemy.orm import Session
+from fastapi import Depends, HTTPException, Query,status
 
 from database import get_db
 from users.models import User
 from notes.models import Note
 from notes.schemas import NoteCreate
+from users.security import get_current_user
 
 def get_note_or_404(
         note_id: int,
-        db: Annotated[Session, Depends(get_db)]
+        db: Annotated[Session, Depends(get_db)],
+        user: Annotated[User, Depends(get_current_user)],
 ) -> Note:
-    note = db.query(Note).filter(Note.id == note_id).first()
+    note = db.query(Note).filter(Note.id == note_id, Note.owner_id == user.id).first()
 
     if note is None:
-        raise HTTPException(status_code=404, detail="Book not found")
+        raise HTTPException(status_code=404, detail="Note not found")
     return note
 
 def create_note(
@@ -33,16 +35,23 @@ def create_note(
 
 def delete_note(
         note: Note,
-        db: Session
+        db: Session,
+        user: User,
 ) -> None:
+    if int(note.owner_id) != user.id:
+        raise HTTPException(status_code=status.HTTP_204_NO_CONTENT, detail="Note not found")
     db.delete(note)
     db.commit() 
 
 def update_note(
         note: Note,
         updated_note: NoteCreate,
+        user: User,
         db: Session,
 ) -> Note:
+    if int(note.owner_id) != user.id:
+        raise HTTPException(status_code=status.HTTP_204_NO_CONTENT, detail="Note not found")
+    
     for field, value in updated_note.model_dump().items():
         setattr(note, field, value)
     
@@ -52,8 +61,12 @@ def update_note(
 
 def toggle_archive(
         note: Note,
+        user: User,
         db: Session,
 ) -> Note:
+    if int(note.owner_id) != user.id:
+        raise HTTPException(status_code=status.HTTP_204_NO_CONTENT, detail="Note not found")
+    
     note.is_archived = not bool(note.is_archived)
     db.commit()
     db.refresh(note)
@@ -61,17 +74,19 @@ def toggle_archive(
 
 def filter_notes(
         db: Annotated[Session, Depends(get_db)],
+        user: Annotated[User, Depends(get_current_user)],
         title: str = Query(default=""),
         skip: int = Query(default=0, ge=0),
-        limit: int = Query(default=10, ge=1, le=20)
+        limit: int = Query(default=10, ge=1, le=20),
+        
 ) -> list[Note]:
     
     query = db.query(Note)
 
     if title is not None:
-        query = query.filter(Note.title.ilike(f"%{title}%"))
+        query = query.filter(Note.title.ilike(f"%{title}%"), Note.owner_id == user.id)
     
     return query.offset(skip).limit(limit).all()
 
-def get_no_of_books(db: Session):
+def get_no_of_notes(db: Session):
     return db.query(Note).count()
